[MINOR] Move T* aliases to ignite-commons
