+++ modules/commons/src/main/java/org/apache/ignite/internal/util/typedef/T1.java
@@ -18,38 +18,33 @@
 package org.apache.ignite.internal.util.typedef;
 
 import java.io.Externalizable;
-import org.apache.ignite.internal.util.lang.GridFunc;
 import org.apache.ignite.internal.util.lang.GridTuple;
-import org.apache.ignite.internal.util.lang.GridTuple5;
 
 /**
- * Defines {@code alias} for {@link GridTuple5} by extending it. Since Java doesn't provide type aliases
+ * Defines {@code alias} for {@link GridTuple} by extending it. Since Java doesn't provide type aliases
  * (like Scala, for example) we resort to these types of measures. This is intended to provide for more
- * concise code in cases when readability won't be sacrificed. For more information see {@link GridTuple5}.
+ * concise code in cases when readability won't be sacrificed. For more information see {@link GridTuple}.
+ * @param <V> Type of the free variable.
  * @see GridFunc
  * @see GridTuple
  */
-public class T5<V1, V2, V3, V4, V5> extends GridTuple5<V1, V2, V3, V4, V5> {
+public class T1<V> extends GridTuple<V> {
     /** */
     private static final long serialVersionUID = 0L;
 
     /**
      * Empty constructor required by {@link Externalizable}.
      */
-    public T5() {
+    public T1() {
         // No-op.
     }
 
     /**
-     * Fully initializes this tuple.
+     * Constructs mutable object with given value.
      *
-     * @param v1 First value.
-     * @param v2 Second value.
-     * @param v3 Third value.
-     * @param v4 Forth value.
-     * @param v5 Fifth value.
+     * @param val Wrapped value.
      */
-    public T5(V1 v1, V2 v2, V3 v3, V4 v4, V5 v5) {
-        super(v1, v2, v3, v4, v5);
+    public T1(V val) {
+        super(val);
     }
 }
